+++ tests/test_style.py
@@ -64,13 +64,6 @@ def test_docformatter():
     assert ['*.py'] == styler.default_extensions()
 
 
-def test_docformatter():
-    styler = zazu.plugins.docformatter_styler.DocformatterStyler()
-    ret = styler.style_string('def foo ():\n"""doc"""\n  pass')
-    assert ret == 'def foo ():\n"""doc"""\n  pass'
-    assert ['*.py'] == styler.default_extensions()
-
-
 def test_goimports(mocker):
     mocker.patch('zazu.util.check_popen', return_value='bar')
     styler = zazu.plugins.goimports_styler.GoimportsStyler(options=['-U'])